requirements.txt
python-can>=4.0.0
colorama>=0.4.6
